requirements.txt
Flask
requests
python-telegram-bot==13.19